Renamed `future_work` dir to `future`
